feat: add technical review report and fix identified bugs
- Add TECHNICAL_REVIEW.md: comprehensive analysis of all modules
- Fix draw_paths_on_pitch: move return outside loop (P0 bug)
- Fix BallTracker.update: skip buffer append on empty detections (P0 bug)
- Fix run_team_classification: guard against empty-team crash (P0 bug)
- Fix TeamClassifier: add random_state=42 to UMAP and KMeans
- Fix FrameState.video_ts: add default None
- Fix GameStateManager: add max_events param, use deque for events
- Fix callbacks: replace silent except with logger.exception
- Fix main.py: add --no_display flag, compute video_ts from fps
- Fix main.py: correct --mode argparse type, add DEFAULT_FPS constant
- Fix ViewTransformer: use RANSAC in findHomography
- Fix SoccerPitchConfiguration.vertices: correct type annotation

+++ tracking/configs/soccer.py
@@ -14,7 +14,7 @@ class SoccerPitchConfiguration:
     penalty_spot_distance: int = 1100  # [cm]
 
     @property
-    def vertices(self) -> List[Tuple[int, int]]:
+    def vertices(self) -> List[Tuple[float, float]]:
         return [
             (0, 0),  # 1
             (0, (self.width - self.penalty_box_width) / 2),  # 2
